requirements.txt
absl-py==0.9.0
alabaster==0.7.11
altgraph==0.16.1
anaconda-client==1.7.2
anaconda-navigator==1.9.6
astor==0.8.0
astroid==2.0.2
attrs==19.3.0
backcall==0.2.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.12
botocore==1.17.12
Brotli==1.0.7
brotlipy==0.7.0
CacheControl==0.12.5
cachetools==4.1.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
conda==4.9.2
conda-build==3.17.6
conda-package-handling==1.7.0
conda-verify==3.1.1
cryptography==2.9.2
cryptography-vectors==2.3
cycler==0.10.0
Cython==0.28.5
dash==1.13.4
dash-bootstrap-components==0.10.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
decorator==4.4.2
defusedxml==0.6.0
dlib==19.21.1
docutils==0.15.2
EasyProcess==0.2.7
entrypoints==0.3
filelock==3.0.12
firebase-admin==2.15.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gast==0.2.2
glob2==0.7
google==3.0.0
google-api-core==1.7.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-images-download==2.4.2
google-pasta==0.2.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.27.2
gym==0.17.2
h5py==2.8.0
idna==2.10
imagecodecs==2020.2.18
imageio==2.9.0
imageio-ffmpeg==0.4.3
ImageScraper==2.0.7
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.1.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
libarchive-c==2.9
lime==0.2.0.0
lxml==4.6.2
macholib==1.11
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
npm==0.1.1
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.1
opt-einsum==3.1.0
optional-django==0.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
playsound==1.2.2
plotly==4.9.0
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycosat==0.6.3
pycparser==2.20
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
PyInstaller==3.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyscreenshot==0.5.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacremoses==0.0.41
scikit-image==0.14.2
scikit-learn==0.24.1
scipy==1.5.0
selenium==3.14.0
Send2Trash==1.5.0
sentencepiece==0.1.91
setproctitle==1.2.2
SimplePool==0.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.5.7
torch==1.3.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
transformers==2.1.1
twilio==6.26.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.1.0
